Initial Push of Filesystem for Entire Ecosystem
This upload contains the file directory for TRMS: The Race Management System, a suite of tools to manage races. This includes TRDS: The Race Data Solution - contains the backend databases, statistics, exports, etc. that will manage both the registration, the race itself and the post race information, TRRS: The Race Registration Solution - which connects to the backend databases to handle registration over the web, the current TRTS: The Race Timing Solution module - released as version 1.0, and TRWS; The Race Web Solution - the UX for the admin to manage post race duties, and for runners to see their results.